Refactor : 메인Desktop기술블로그 컴포넌트 분리 및 모바일 분기처리
