+++ components/features/main/dynamicTechBlogComponent.tsx
@@ -1,31 +1,21 @@
 import React from 'react';
 
-import Image from 'next/image';
-import Link from 'next/link';
-
-import { useQueryClient } from '@tanstack/react-query';
+import dynamic from 'next/dynamic';
 
 import NoMyInfoData from '@pages/myinfo/components/NoMyInfoData';
-import { usePostBookmarkStatus } from '@pages/techblog/api/usePostBookmarkStatus';
-import { ArticleViewBtn } from '@pages/techblog/components/techDetailCardSubComponent';
-import { TechContent, TechInfo } from '@pages/techblog/components/techSubComponents';
 import { TechCardProps } from '@pages/techblog/types/techBlogType';
-
-import { useToastVisibleStore } from '@stores/toastVisibleStore';
 
 import useIsMobile from '@hooks/useIsMobile';
 import { useObserver } from '@hooks/useObserver';
 
 import { MainTechSkeletonList } from '@components/common/skeleton/techBlogSkeleton';
 
-import bookmarkActive from '@public/image/techblog/bookmarkActive.svg';
-import bookmarkNonActive from '@public/image/techblog/bookmarkNonActive.svg';
-
-import { ROUTES } from '@/constants/routes';
 import { TechInfiniteDataType } from '@/types/infiniteQueryType';
 
-import TechBlogImg from '../techblog/techBlogImg';
+import DesktopMainTechCard from '../techblog/desktopMainTechCard';
 import GradientDiv from './gradientDiv';
+
+const DynamicTechCard = dynamic(() => import('@/pages/techblog/components/techCard'));
 
 export default function DynamicTechBlogComponent({
   skeletonCnt,
@@ -42,9 +32,6 @@ export default function DynamicTechBlogComponent({
 }) {
   const isMobile = useIsMobile();
 
-  const queryClient = useQueryClient();
-  const { setToastVisible } = useToastVisibleStore();
-
   const { techBlogData, isFetchingNextPage, hasNextPage, status, onIntersect } = data;
 
   const SCROLL_CLASS = isMobile ? '' : 'relative overflow-y-scroll scrollbar-hide max-h-[50rem]';
@@ -53,29 +40,6 @@ export default function DynamicTechBlogComponent({
     target: bottomDiv,
     onIntersect,
   });
-
-  const { mutate: bookmarkMutation } = usePostBookmarkStatus();
-
-  const handleBookmarkClick = ({
-    id,
-    isBookmarkActive,
-  }: {
-    id: number;
-    isBookmarkActive: boolean;
-  }) => {
-    bookmarkMutation(
-      {
-        techArticleId: id,
-        status: !isBookmarkActive,
-      },
-      {
-        onSuccess: async () => {
-          await queryClient.invalidateQueries({ queryKey: ['techBlogBookmark'] });
-          setToastVisible('북마크에서 삭제했어요');
-        },
-      },
-    );
-  };
 
   const getStatusComponent = () => {
     switch (status) {
@@ -90,73 +54,15 @@ export default function DynamicTechBlogComponent({
             <div className={isScroll ? SCROLL_CLASS : ''}>
               {techBlogData?.pages?.map((group, index) => (
                 <React.Fragment key={index}>
-                  {group?.data.content.map(
-                    ({
-                      id,
-                      thumbnailUrl,
-                      techArticleUrl,
-                      title,
-                      isBookmarked,
-                      author,
-                      regDate,
-                      company,
-                      contents,
-                    }: TechCardProps) => (
-                      <div
-                        key={id}
-                        className='grid grid-flow-col border-white gap-[3.2rem] text-white py-[2.8rem] border-b border-b-gray1 border-solid select-none'
-                      >
-                        <div>
-                          <TechBlogImg
-                            id={id}
-                            thumbnailUrl={thumbnailUrl}
-                            rounded='rounded-[0.8rem]'
-                            size='small'
-                          />
-
-                          <ArticleViewBtn
-                            techArticleUrl={techArticleUrl}
-                            fontSize='c1'
-                            textIconGap={'mr-[0.8rem]'}
-                            paddingY='pt-[1.6rem]'
-                            iconSize='w-[6px] h-[20px]'
-                          />
-                        </div>
-                        <div>
-                          <div className='flex items-center justify-between border-white'>
-                            <Link href={`${ROUTES.TECH_BLOG}/${id}`}>
-                              <p className='font-bold st2 py-[0.7rem]'>{title}</p>
-                            </Link>
-                            {type === 'myinfo' && (
-                              <Image
-                                src={isBookmarked ? bookmarkActive : bookmarkNonActive}
-                                width={15}
-                                height={16}
-                                alt='북마크활성아이콘'
-                                className='cursor-pointer'
-                                onClick={() =>
-                                  handleBookmarkClick({
-                                    id,
-                                    isBookmarkActive: isBookmarked,
-                                  })
-                                }
-                              />
-                            )}
-                          </div>
-                          <TechInfo
-                            type='main'
-                            author={author}
-                            date={regDate}
-                            company={company?.name}
-                            companyId={id}
-                          />
-                          <Link href={`${ROUTES.TECH_BLOG}/${id}`}>
-                            <TechContent content={contents} maxLines={4} className='mr-4' />
-                          </Link>
-                        </div>
-                      </div>
-                    ),
-                  )}
+                  {group?.data.content.map((data: TechCardProps) => (
+                    <>
+                      {isMobile ? (
+                        <DynamicTechCard key={data.id} techData={data} />
+                      ) : (
+                        <DesktopMainTechCard key={data.id} techData={data} type={type} />
+                      )}
+                    </>
+                  ))}
                 </React.Fragment>
               ))}
             </div>
